+++ spark-bigquery-dsv2/spark-3.4-bigquery/src/test/java/com/google/cloud/spark/bigquery/integration/Spark34DirectWriteIntegrationTest.java
@@ -15,7 +15,25 @@
  */
 package com.google.cloud.spark.bigquery.integration;
 
+import static com.google.common.truth.Truth.assertThat;
+
+import com.google.cloud.bigquery.BigQuery;
+import com.google.cloud.bigquery.LegacySQLTypeName;
+import com.google.cloud.bigquery.QueryJobConfiguration;
+import com.google.cloud.bigquery.TableResult;
 import com.google.cloud.spark.bigquery.SparkBigQueryConfig;
+import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
+import org.apache.spark.sql.Dataset;
+import org.apache.spark.sql.Row;
+import org.apache.spark.sql.RowFactory;
+import org.apache.spark.sql.SaveMode;
+import org.apache.spark.sql.types.DataTypes;
+import org.apache.spark.sql.types.Metadata;
+import org.apache.spark.sql.types.StructField;
+import org.apache.spark.sql.types.StructType;
+import org.junit.Test;
 
 public class Spark34DirectWriteIntegrationTest extends WriteIntegrationTestBase {
 
@@ -24,4 +42,31 @@ public class Spark34DirectWriteIntegrationTest extends WriteIntegrationTestBase 
   }
 
   // tests from superclass
+  @Test
+  public void testTimestampNTZWriteToBigQuery() throws InterruptedException {
+    LocalDateTime time = LocalDateTime.of(2023, 9, 1, 12, 23, 34, 268543 * 1000);
+    List<Row> rows = Arrays.asList(RowFactory.create(time));
+    Dataset<Row> df =
+        spark.createDataFrame(
+            rows,
+            new StructType(
+                new StructField[] {
+                  StructField.apply("foo", DataTypes.TimestampNTZType, true, Metadata.empty())
+                }));
+    String table = testDataset.toString() + "." + testTable;
+    df.write()
+        .format("bigquery")
+        .mode(SaveMode.Overwrite)
+        .option("table", table)
+        .option("writeMethod", SparkBigQueryConfig.WriteMethod.DIRECT.toString())
+        .save();
+    BigQuery bigQuery = IntegrationTestUtils.getBigquery();
+    TableResult result =
+        bigQuery.query(
+            QueryJobConfiguration.of(String.format("Select foo from %s", fullTableName())));
+    assertThat(result.getSchema().getFields().get(0).getType())
+        .isEqualTo(LegacySQLTypeName.DATETIME);
+    assertThat(result.streamValues().findFirst().get().get(0).getValue())
+        .isEqualTo("2023-09-01T12:23:34.268543");
+  }
 }